refactor(ui): colocate memory, budgets, and access-groups into _components/
Continues the colocation follow-up (after api-reference #32690), applying the
same convention to three more segments whose local component folders are
entirely self-contained: rename each `<segment>/components/` to `_components/`
(private to the segment, matching Next's `_` route-exclusion) and repoint the
one `page.tsx` import.

Each folder is imported only by its own page (verified zero external importers
across src, tests, and e2e_tests), and every `@/components` import in these
segments is genuinely shared infrastructure (common_components, shared/*,
networking, molecules), so nothing is drained out of `@/components`. Pure file
relocation, no logic change.

The grandfathered lint suppressions on budgets (@tremor/react) and memory
(react-hooks/set-state-in-effect) are re-keyed to the new paths; the
suppression counts are unchanged.
