+++ ui/litellm-dashboard/src/app/(dashboard)/budgets/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import BudgetPanel from "./components/budget_panel";
+import BudgetPanel from "./_components/budget_panel";
 import useAuthorized from "@/app/(dashboard)/hooks/useAuthorized";
 
 export default function Budgets() {
